Add basic e2e tests for repository tags
